1.0.0-alpha (#17)
1.0.0-alpha版本发布,完善需求、缺陷、流水线、用例、系统工具功能

+++ windy-domain/src/main/java/com/zj/domain/entity/bo/demand/DemandBO.java
@@ -73,7 +73,7 @@ public class DemandBO {
   /**
    * 工作量
    */
-  private Double workload;
+  private Integer workload;
 
   /**
    * 需求标签
